fetching fee details from fee api (#1255)
* removing Address history complete section and address history for person staying same address for 5 year also fix address details screen bug

* PRL-5116

* PRL-5116
